+++ internal/entity/delivery_entity.go
@@ -4,8 +4,8 @@ import "time"
 
 type Delivery struct {
 	ID         uint64    `gorm:"primary_key;column:id;autoIncrement"`
-	Cost       int       `gorm:"column:cost"`
-	Distance   int       `gorm:"column:distance"`
+	Cost       float32   `gorm:"column:cost"`
+	Distance   float32   `gorm:"column:distance"`
 	Created_At time.Time `gorm:"column:created_at;autoCreateTime;<-:create"`
 	Updated_At time.Time `gorm:"column:updated_at;autoCreateTime;autoUpdateTime"`
 }
