+++ internal/entity/product_entity.go
@@ -8,7 +8,7 @@ type Product struct {
 	CategoryId  uint64    `gorm:"column:category_id"`
 	Name        string    `gorm:"column:name"`
 	Description string    `gorm:"column:description"`
-	Price       int       `gorm:"column:price"`
+	Price       float32   `gorm:"column:price"`
 	Stock       int       `gorm:"column:stock"`
 	Created_At  time.Time `gorm:"column:created_at;autoCreateTime;<-:create"`
 	Updated_At  time.Time `gorm:"column:updated_at;autoCreateTime;autoUpdateTime"`
