+++ internal/model/discount_model.go
@@ -9,7 +9,7 @@ type DiscountResponse struct {
 	Name        string              `json:"name,omitempty"`
 	Description string              `json:"description,omitempty"`
 	Code        string              `json:"code,omitempty"`
-	Value       int                 `json:"value,omitempty"`
+	Value       float32             `json:"value,omitempty"`
 	Type        helper.DiscountType `json:"type,omitempty"`
 	Start       string              `json:"start,omitempty"`
 	End         string              `json:"end,omitempty"`
@@ -22,7 +22,7 @@ type CreateDiscountRequest struct {
 	Name        string              `json:"name" validate:"required,max=100"`
 	Description string              `json:"description" validate:"required"`
 	Code        string              `json:"code" validate:"required,max=100"`
-	Value       int                 `json:"value" validate:"required"`
+	Value       float32             `json:"value" validate:"required"`
 	Type        helper.DiscountType `json:"type" validate:"required"`
 	Start       string              `json:"start" validate:"required"`
 	End         string              `json:"end" validate:"required"`
@@ -38,7 +38,7 @@ type UpdateDiscountRequest struct {
 	Name        string              `json:"name" validate:"required,max=100"`
 	Description string              `json:"description" validate:"required"`
 	Code        string              `json:"code" validate:"required,max=100"`
-	Value       int                 `json:"value" validate:"required"`
+	Value       float32             `json:"value" validate:"required"`
 	Type        helper.DiscountType `json:"type" validate:"required"`
 	Start       string              `json:"start" validate:"required"`
 	End         string              `json:"end" validate:"required"`
